+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/DeployJob.java
@@ -79,7 +79,7 @@ public class DeployJob extends WorkspaceJob {
   private final DefaultDeployConfiguration deployConfiguration;
   private final boolean includeOptionalConfigurationFiles;
   private final CollectingLineListener errorCollectingLineListener;
-  private final DeployEnvironmentDelegate environmentDelegate;
+  private final StagingDelegate stager;
 
   /**
    * @param workDirectory temporary work directory the job can safely use (e.g., for creating and
@@ -100,7 +100,7 @@ public class DeployJob extends WorkspaceJob {
       ProcessOutputLineListener stderrLineListener,
       DefaultDeployConfiguration deployConfiguration,
       boolean includeOptionalConfigurationFiles,
-      DeployEnvironmentDelegate environmentDelegate) {
+      StagingDelegate stager) {
     super(Messages.getString("deploy.job.name")); //$NON-NLS-1$
     this.project = project;
     this.credential = credential;
@@ -109,7 +109,7 @@ public class DeployJob extends WorkspaceJob {
     this.stderrLineListener = stderrLineListener;
     this.deployConfiguration = deployConfiguration;
     this.includeOptionalConfigurationFiles = includeOptionalConfigurationFiles;
-    this.environmentDelegate = environmentDelegate;
+    this.stager = stager;
     deployStdoutLineListener = new StringBuilderProcessOutputLineListener();
     errorCollectingLineListener =
         new CollectingLineListener(new Predicate<String>() {
@@ -176,7 +176,7 @@ public class DeployJob extends WorkspaceJob {
     try {
       getJobManager().beginRule(project, null /* not worth a monitor */);
       IPath safeWorkDirectory = workDirectory.append(SAFE_STAGING_WORK_DIRECTORY_NAME);
-      IStatus status = environmentDelegate.stage(
+      IStatus status = stager.stage(
           project, stagingDirectory, safeWorkDirectory, cloudSdk, monitor);
       if (stagingExitListener.getExitStatus() != Status.OK_STATUS) {
         return stagingExitListener.getExitStatus();
@@ -196,8 +196,7 @@ public class DeployJob extends WorkspaceJob {
 
     IPath optionalConfigurationFilesDirectory = null;
     if (includeOptionalConfigurationFiles) {
-      optionalConfigurationFilesDirectory =
-          environmentDelegate.getOptionalConfigurationFilesDirectory();
+      optionalConfigurationFilesDirectory = stager.getOptionalConfigurationFilesDirectory();
     }
 
     new AppEngineProjectDeployer().deploy(stagingDirectory, cloudSdk, deployConfiguration,
