+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/flex/FlexStagingDelegate.java
@@ -17,7 +17,7 @@
 package com.google.cloud.tools.eclipse.appengine.deploy.flex;
 
 import com.google.cloud.tools.appengine.cloudsdk.CloudSdk;
-import com.google.cloud.tools.eclipse.appengine.deploy.DeployEnvironmentDelegate;
+import com.google.cloud.tools.eclipse.appengine.deploy.StagingDelegate;
 import com.google.cloud.tools.eclipse.appengine.deploy.DeployStaging;
 import com.google.cloud.tools.eclipse.appengine.deploy.WarPublisher;
 import org.eclipse.core.resources.IProject;
@@ -28,11 +28,11 @@ import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.SubMonitor;
 
-public class FlexDeployEnvironmentDelegate implements DeployEnvironmentDelegate {
+public class FlexStagingDelegate implements StagingDelegate {
 
   private final IPath appEngineDirectory;
 
-  FlexDeployEnvironmentDelegate(IPath appEngineDirectory) {
+  FlexStagingDelegate(IPath appEngineDirectory) {
     this.appEngineDirectory = appEngineDirectory;
   }
